feat: 添加了最热排行榜和最新排行榜的后端逻辑（暂未写数据库处理）
	renamed:    file/src/main/java/com/scut/file/config/FeignConfig.java -> board/src/main/java/com/scut/board/config/FeignConfig.java~merged
	modified:   board/src/main/java/com/scut/board/controller/ViewBoardController.java
	new file:   board/src/main/java/com/scut/board/entity/GameForBoardEntity.java
	new file:   board/src/main/java/com/scut/board/mapper/ViewBoardMapper.java
	new file:   board/src/main/java/com/scut/board/mapper/ViewBoardMapperImpl.java
	new file:   board/src/main/java/com/scut/board/service/ViewBoardService.java
	new file:   board/src/main/java/com/scut/board/service/ViewBoardServiceImpl.java
	modified:   common/src/main/java/com/scut/common/dto/response/GameForBoardDto.java

+++ common/src/main/java/com/scut/common/dto/response/GameForBoardDto.java
@@ -15,5 +15,5 @@ public class GameForBoardDto {
     String cover; //封面地址
     String logo; //logo地址
     String classify; //游戏分类
-    long score; //评分
+    float score; //评分
 }
